requirements.txt
fastapi
uvicorn
python-dotenv
httpx
gdown
torch
numpy
pillow
opencv-python
timm
torchvision
